add original version of unit test

+++ database/migrations/2022_07_05_201022_create_samples_table.php
@@ -6,14 +6,13 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     public function up(): void
     {
         Schema::create('samples', static function (Blueprint $table) {
             $table->id();
             $table->string('device_id');
-            $table->tinyInteger('temp');
+            $table->float('temp');
             $table->timestamp('created_at')->useCurrent();
         });
     }
